requirements.txt
numpy==1.16.2
pysftp==0.2.9
pandas==0.24.2
netcdf4==1.5.0
psycopg2==2.7.7
gdal==2.4.1
geopandas==0.5.1
fastText==0.9.1
twitterapi==2.5.9
simplejson==3.16.0
xarray==0.12.1
tensorflow-gpu==1.13.1
pytorch==1.0.1
dask==2.3.0
matplotlib==3.1.1
openpyxl==1.0.1
xlrd==1.2.0